+++ src/cosmic/src/const_bse.h
@@ -33,9 +33,10 @@
       COMMON /CEVARS/ qcrit_array,alpha1,lambdaf
       REAL*8 bconst,CK
       COMMON /MAGVARS/ bconst,CK
-      INTEGER kickflag
+      INTEGER kickflag,fryer_mass_limit
       REAL*8 sigma,sigmadiv,bhsigmafrac,pisn,mxns
       REAL*8 polar_kick_angle
+      REAL*8 ppi_co_shift,ppi_extra_ml
       REAL*8 ecsn,ecsn_mlow,bhspinmag,rembar_massloss
       REAL*8 mm_mu_ns, mm_mu_bh, maltsev_fallback,maltsev_pf_prob
       REAL*8 natal_kick_array(2,5)
@@ -44,7 +45,8 @@
      &            polar_kick_angle,pisn,ecsn,ecsn_mlow,
      &            bhspinmag,mxns,rembar_massloss,
      &            mc_he,mc_co,mm_mu_ns,mm_mu_bh,maltsev_fallback,
-     &            maltsev_pf_prob,kickflag
+     &            maltsev_pf_prob,kickflag,fryer_mass_limit,
+     &            ppi_co_shift,ppi_extra_ml
       REAL*8 fprimc_array(16)
       COMMON /TIDALVARS/ fprimc_array
       REAL*8 rejuv_fac