+++ changelog.md
@@ -71,7 +71,34 @@ See the discussed changes in our previous releases here: https://github.com/COSM
  - Add `remnantflag=6`: The Maltsev+25 and Willcox+25 prescription for remnant masses. This can be tuned with `maltsev_mode` and `maltsev_fallback`
  - (Docs) Any tutorial that uses a plain BSEDict now uses one that is drawn from the cosmic-settings.json file to avoid missing changes and saves us updating each docs page
 
+
 ## 3.7.1
  - Bug fix [[#729](https://github.com/COSMIC-PopSynth/COSMIC/issues/729)]: ensure disruptions are logged with ``evol_type==11`` when SNe occur during CEs
  - Bug fix [[#725](https://github.com/COSMIC-PopSynth/COSMIC/issues/725)]: set ``tb=sep=0`` for cases where AIC caused a WD to explode and leave behind no remnant (``evolve_type==9``)
  - Bug fix [[#724](https://github.com/COSMIC-PopSynth/COSMIC/issues/724)]: remove bug where ``kstar=15`` was assigned its previous epoch mass after merging during a CE and the merger product goes SN
+
+## 3.7.2
+This release contains _several_ fixes to how CO core masses/remnant masses are handled. It also adds a new PISN prescription and windflag.
+
+- Fixes:
+    - Update ``mc_co`` and ``mc_he`` after adjusting ``mc = mcmax`` in ``hrdiag.f`` for stripped stars. This can be a fairly significant change, up to ~2 Msun. (Used to actually be up to 10 Msun because we added HeMS core mass growth)
+    - Make ``evolv2.f`` actually log PISN (they get set to ``kw = 15`` so were missed previously)
+    - Change ``loop = 40000`` (up from 20,000), otherwise 70 Msun stars run out of loops with new mass modifier
+    - Changed all existing ``pisn`` prescriptions to use ``mc_tot`` instead of ``mcbagb`` for conditions
+    - Update ``mc = mt`` for **ALL** remnants instead of just for certain flags
+    - Only apply conversion from baryonic to gravitational mass once. Previous code **APPLIED IT TWICE!!** for core-collapse BHs! Create subroutine ``baryonic_to_gravitational_mass``, which is called throughout
+    - Ensure stellar type changes during hrdiag are logged to the bpp before SN occurs
+
+- Additions/changes:
+    - Linearly grow the core mass of HeMS stars from 0 to the value at the start of HeHG and check for core-collapse for all He stars (not just HeHG and HeGB)
+    - Add ``dt_mass_modifier`` argument to ``Evolve.evolve()``. This lets you specify a list of tuples ``(m_low, m_high, mod)``, which multiplies ``pts1/pts2/pts3`` by ``mod`` for stars with primary star (only care about primary, conservative) ZAMS masses between ``m_low <= m_ZAMS < m_high``
+    - Added a ``windflag = -1`` option to turn off stellar winds entirely
+    - Added ``pisn = -4`` for Renzo+22/Hendriks+23 prescription for PPISN mass loss
+        - With options ``ppi_co_shift`` and ``ppi_extral_ml``
+
+- Code cleanup:
+    - ``assign_remnant`` no longer takes ``mc_tot`` as a parameter, just get it from the common block
+    - WD chandrasekhar check use ``mc_co`` instead of ``mc``
+    - Remnant flags 0-4 inclusive use ``mc_co`` instead of ``mc``
+    - Got rid of ``mcx`` in ``assign_remnant`` in favour of clearer ``m_proto`` and ``m_FeNi`` to match the papers
+    - [Very minor] Fryer Rapid was using <= instead of < everywhere